comperank: Ranking and Rating Methods for Competition Results

comperank provides tools for computing ranking and rating based on competition results. It is tightly connected to its data infrastructure package comperes. Basic knowledge about creating valid competition results and Head-to-Head expressions with comperes is needed in order to efficiently use comperank.

Details

comperank leverages the tidyverse ecosystem of R packages. Among other things, it means that the main output format is tibble.

comperank gets inspiration from the book "Who's #1" by Langville and Meyer. It provides functionality for the following rating algorithms:

Algorithms with fixed Head-to-Head structure:

Simplified Massey method with rate_massey() and rank_massey().

Simplified Colley method with rate_colley() and rank_colley().

Algorithms with variable Head-to-Head structure:

Keener method with rate_keener() and rank_keener().

Markov method with rate_markov() and rank_markov().

Offense-Defense method with rate_od() and rank_od().

Algorithms with iterative nature:

General Iterative ratings with rate_iterative(), rank_iterative(), and add_iterative_ratings().

Elo ratings with rate_elo(), rank_elo(), and add_elo_ratings().

comperank also offers data sets describing professional snooker in seasons 2016/2017 and 2017/2018. See snooker_events, snooker_players, snooker_matches.

To learn more about comperank browse vignettes with browseVignettes(package = "comperank").

See also