comperes: Manage Competition Results

comperes offers a set of tools for storing and managing competition results. Competition is understood as a set of games (abstract event) in which players (abstract entity) gain some abstract scores. The most natural example is sport results, however not the only one. For example, product rating can be considered as a competition between products as "players". Here a "game" is a customer that reviews a set of products by rating them with numerical "score" (stars, points, etc.).

Details

This package provides the following functionality:

Store and convert competition results:

In long format as a tibble with one row per game-player pair.

In wide format as a tibble with one row per game with fixed amount of players.

Summarise:

Compute item summaries with functions using dplyr's grammar of data manipulation.

Compute and join item summaries to data for easy transformation.

Use common item summary functions with rlang's unquoting mechanism.

Compute Head-to-Head values (a summary statistic of direct confrontation between two players) with functions also using dplyr's grammar:

Store output in long format as a tibble with one row per pair of players.

Store output in matrix format as a matrix with rows and columns describing players and entries - Head-to-Head values.

Use common Head-to-Head functions with rlang's unquoting mechanism.

To learn more about comperes browse vignettes with browseVignettes(package = "comperes").

See also

Useful links:

https://github.com/echasnovski/comperes

Report bugs at https://github.com/echasnovski/comperes/issues