Snooker matches

Data set describing snooker matches in seasons 2016/2017 and 2017/2018.

snooker_matches

Format

A tibble with one row per match and the following columns:

id <int> : Match identifier in snooker.org database.

eventId <int> : Match's event identifier (taken from id column of snooker_events)

round <int> : Round number of event in which match was played. Usually event's structure is organized in rounds: sets of matches with roughly "the same importance". Usually the more round number the "more important" matches are played. However, there are many exceptions.

player1Id <int> : Identifier of first player in match (taken from id column of snooker_players).

score1 <int> : Number of won frames (individual games) by first player.

walkover1 <lgl> : Whether the win of first player was scored by the technical reasons.

player2Id <int> : Identifier of second player in match (taken from id column of snooker_players).

score2 <int> : Number of won frames (individual games) by second player.

walkover2 <lgl> : Whether the win of second player was scored by the technical reasons.

winnerId <int> : Identifier of match's winner (taken from either player1Id or player2Id columns).

startDate <dttm> : Time at which match started.

endDate <dttm> : Time at which match ended.

scheduledDate <dttm> : Time at which match was scheduled to start.

frameScores <chr> : Scores of players in frames. Usually is missing, present only for important matches.

Details

Data is taken from snooker.org (http://www.snooker.org/) API.

Matches are present only for tracked snooker events.

See also