Snooker players

Data set describing snooker players in seasons 2016/2017 and 2017/2018.

snooker_players

Format

A tibble with one row per player and the following columns:

id <int> : Player identifier in snooker.org database (used in player1Id, player2Id and winnerId columns of snooker_matches).

name <chr> : Player full name.

nationality <chr> : Player nationality.

sex <chr> : Player gender ("F" for female, "M" for male, and "Unknown").

born <dttm> : Player date of birth.

status <chr> : Player status in season 2017/2018 ("pro" for professional, "ama" for amateur).

Details

Data is taken from snooker.org (http://www.snooker.org/) API.

Data is present only for players who played at least one game in tracked snooker events in seasons 2016/2017 and 2017/2018.

See also