Common item summary functions

List of commonly used functions for summarising competition results.

summary_funs

Format

An object of class list of length 8.

Details

summary_funs is a named list of expressions representing commonly used expressions of summary functions for summarising competition results with summarise_item(). Names of the elements will be used as summary names. It is designed primarily to be used with long format of competition results. To use them inside summarise_item() use unquoting mechanism from rlang package.

Currently present functions:

min_score - min(score).

max_score - max(score).

mean_score - mean(score).

median_score - median(score).

sd_score - sd(score).

sum_score - sum(score).

num_games - length(unique(game)).

num_players - length(unique(player)).

Note that it is generally better to subset summary_funs using names rather than indices because the order of elements might change in future versions.

See also

Compute item summary, Join item summary

Examples

ncaa2005 %>% summarise_game(!!! summary_funs, .prefix = "game_")

#> # A tibble: 10 x 9
#>     game game_min_score game_max_score game_mean_score game_median_score
#>    <int>          <dbl>          <dbl>           <dbl>             <dbl>
#>  1     1             7.            52.           29.5              29.5 
#>  2     2            21.            24.           22.5              22.5 
#>  3     3             7.            38.           22.5              22.5 
#>  4     4             0.            45.           22.5              22.5 
#>  5     5            16.            34.           25.0              25.0 
#>  6     6            17.            25.           21.0              21.0 
#>  7     7             7.            27.           17.0              17.0 
#>  8     8             5.             7.            6.00              6.00
#>  9     9             3.            30.           16.5              16.5 
#> 10    10            14.            52.           33.0              33.0 
#> # ... with 4 more variables: game_sd_score <dbl>, game_sum_score <int>,
#> #   game_num_games <int>, game_num_players <int>